[SWIK-738_social_login] Moved handler code of social stuff in own file. created util.js for shared functions
